+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[UPDATE]** Allow change of `defaultValue` in `SelectField`
 - [...]
 
 # v41.12.0 (08/12/2020)